Time-keeping updates
Enable rtcfile and rtcautotrim options in chrony if there
is a battery-backed RTC. Perhaps could enable it even if
the RTC is not battery powered. It isn't clear in the
chrony docs whether chrony initializes the system clock
from the RTC. If the RTC is not battery backed we don't
want chrony to use it to set the system clock.

Create and enable ublox-gps service to configured a ublox 6Q.