add tinyproxy to connect to cell modem web interface; add usb-modeswitch
Apparently I never added the usb-modeswitch package anywhere, but it should
be a dependency of isfs-daq since it modifies the usb-modeswitch rules for
the cell modems.  Also, add the tinyproxy reverse proxy server and a config
file so it's possible to connect to the modem web interface over an ssh
tunnel.