Put new hostname before localhost.
When using rsyslog to log to a server, rsyslog apparently takes the
first entry on the 127.0.0.1 line as the hostname.  This is usually
localhost, which is not a useful name on the server side.

Previous behaviour was to add the new hostname at the end, and then
remove the old hostname.  It would leave additional names like 'viper'
or 'titan'.

New behaviour completely replaces the 127.0.0.1 line with "hostname
localhost".  All additional names on the line are lost.