Removed rs485 temperature sensor configuration
From Josh:
Over the last 6 mo. all operations VortexDSMs have been retrofitted with a different temperature sensor on COM3(/dev/ttyS2). This sensor is a long-term solution and can be viewed as a heartbeat data generator for all DSMs.
From Catherine:
I decided to remove the line as the comment has the example. Kept the
empty file because there could be future 485/422 sensors.